% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks}
\alias{checks}
\title{Checks the arguments for the GEARS function.}
\usage{
checks(
  DATA,
  forecast.horizon,
  size.rs,
  number.rs,
  glm.family,
  level,
  y.name = NULL,
  y.max.lags = NULL,
  x.names = NULL,
  x.max.lags = NULL,
  x.fixed.names = NULL,
  x.fixed.lags = NULL,
  x.interaction.names = NULL,
  x.interaction.lags = NULL,
  last.obs,
  use.intercept,
  error.measure,
  betas.selection
)
}
\arguments{
\item{DATA}{A data frame or a univariate time series.}

\item{forecast.horizon}{Number of periods for forecasting.}

\item{size.rs}{Number of observations in the rolling sample.}

\item{number.rs}{Number of rolling samples.}

\item{glm.family}{A description of the error distribution to be used in the
model. See \link[stats]{glm} for details.}

\item{level}{Confidence level for prediction intervals. Numeric value
between 0 and 100.}

\item{y.name}{The name of the Y (left-hand side) variable. If NULL (default),
the function creates a temporary name.}

\item{y.max.lags}{A numeric value that gives the maximum number of lags of
the Y (left-hand side) variable (see Details). Can be NULL (default) if
the past values of the Y variable are not included in the model.}

\item{x.names}{List with names of the X (right-hand side) variables that
have a maximum number of lags (see Details). Can be NULL (default) if
univariate model or if your model does not have variables of this type.}

\item{x.max.lags}{List of numeric values that give the maximum number of lags
of the X (right-hand side) variables. Can be NULL (default).}

\item{x.fixed.names}{List with names of the X (right-hand side) variables
that have a fixed number of lags (see Details).
Can be NULL (default) if univariate model or if your model does not have
variables of this type.}

\item{x.fixed.lags}{List of numeric values that give the fixed number of lags
of the variables in x.fixed.names. Can be NULL (default).}

\item{x.interaction.names}{List of character vectors with names of the
variables to be included as interaction terms (see Details).
Can be NULL (default) if your model does not have interaction terms.}

\item{x.interaction.lags}{List of numeric vectors with lags of the
variables to be included as interaction terms. List and numeric vectors
should have the same length as the ones in x.interaction.names.
Can be NULL (default) if your model does not have interaction terms.}

\item{last.obs}{Index number of the last observation to be considered.}

\item{use.intercept}{Possible values \code{c("both", "with", "without")}.}

\item{error.measure}{Possible values
\code{c("mse", "mae", "mase", "smape", "owa")}.}

\item{betas.selection}{Possible values \code{c("last", "average", "both")}.}
}
\value{
Stops the program and returns an error message.
}
\description{
This function checks the arguments and stops the program if encounters an
error. An error message is returned to the user.
}
\keyword{internal}
