% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01data_definitions.R
\docType{data}
\name{commodities_prices}
\alias{commodities_prices}
\title{Primary Commodities Price Data}
\format{
A dataframe with 483 observations, and 6 variables.
   The variables are described in the
   \code{\link{commodities_prices_data_dictionary}}.
}
\source{
\url{https://www.imf.org/~/media/Files/Research/CommodityPrices/Monthly/Table3.ashx}
}
\usage{
commodities_prices
}
\description{
The most recent data from the International Monetary Fund on Primary
Commodities Prices. In particular, data on Beef, Swine (pork) meat,
Poultry (chicken) meat, Maize (corn), and Wheat.
Data is collected monthly, starting on January, 1980.
}
\keyword{datasets}
