% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01data_definitions.R
\docType{data}
\name{commodities_prices_data_dictionary}
\alias{commodities_prices_data_dictionary}
\title{Descriptions of the commoditiesprices data.}
\format{A dataframe with 6 observations of fields, and 8 variables.
   \describe{\item{variable_name}{The variable name as it appears in the
   \code{\link{commodities_prices}} data frame.}
   \item{variable_description}{Detailed description of the contents of each
   variable.}
   \item{data_type}{Variable type: character (char) or numeric (num).}
   \item{units}{Monetary unit.}
   \item{frequency}{Time frequency of the data.}
   \item{source}{Name of the data source.}
   \item{source_link}{URL Link to the data source.}
   \item{access_on}{Date in which the data used to create the
   \code{\link{commodities_prices}} data frame was obtained.}}}
\usage{
commodities_prices_data_dictionary
}
\description{
A data dictonary describing the commoditiesprices data.
}
\keyword{datasets}
