% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseasonalize.R
\name{deseason}
\alias{deseason}
\title{Function to deseasonalize a time series.}
\usage{
deseason(ts.data, ts.frequency, forecast.horizon, alpha.level = 0.05)
}
\arguments{
\item{ts.data}{A ts object (i.e., a vector of time series observations).}

\item{ts.frequency}{The frequency of a ts object.}

\item{forecast.horizon}{Number of periods for forecasting.}

\item{alpha.level}{The alpha level to be used in the test to detect
seasonality. Default is 0.05.}
}
\value{
A list with the deseasonalized data (\code{"deseasonTS"}) and the
    seasonal component (\code{"seasonalComp"}).
}
\description{
This function deseasonalizes a time series using the "multiplicative"
decomposition only.
}
\examples{
# Example using a TS with frequency > 1 (in this case, frequency = 12)
tmpAir <- deseason(
  ts.data          = datasets::AirPassengers,
  ts.frequency     = stats::frequency(datasets::AirPassengers),
  forecast.horizon = 20
)

head(datasets::AirPassengers) # original data
head(tmpAir$deseasonTS)       # deseasonalized series w/o multiplicative
                              # component, not equal to the original
head(tmpAir$deseasonTS * tmpAir$seasonalComp) # this should be equal to the
                                              # original data
# Example using a TS with frequency = 1
tmpUsage <- deseason(
  ts.data          = datasets::WWWusage,
  ts.frequency     = stats::frequency(datasets::WWWusage),
  forecast.horizon = 10
)

head(datasets::WWWusage)  # original data
head(tmpUsage$deseasonTS) # since frequency = 1, results should be the same

}
