% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluationM4.R
\name{evaluationM4}
\alias{evaluationM4}
\title{Function to evaluate the forecasts following the M4 Competition guidelines}
\usage{
evaluationM4(DATA, forecast.list, alpha)
}
\arguments{
\item{DATA}{A ts object (i.e., a vector of time series observations).}

\item{forecast.list}{A list with ts objects.}

\item{alpha}{Significance level}
}
\value{
A list with three values: "smape", "mase", and "owa".
}
\description{
Function to evaluate the forecasts following the M4 Competition guidelines
}
