% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_best.R
\name{fit_best}
\alias{fit_best}
\title{Estimates the Best Equation =================================================}
\usage{
fit_best(
  forecast.lead,
  eq.number,
  sample.number,
  glm.family,
  Y.name,
  all.equations.rhs,
  DF.Fit.Predict,
  ...
)
}
\arguments{
\item{forecast.lead}{Number of the forecast lead.}

\item{eq.number}{Number (position)of the equation in the vector of all
possible equations.}

\item{sample.number}{Number of the specific random sample that will be
evaluated.}

\item{glm.family}{A description of the error distribution to be used in the
model. Inherits from \code{gears}.}

\item{Y.name}{Name of the y variable (variable to be forecasted) created by
\code{gears}.}

\item{all.equations.rhs}{Vector with all possible combinations of right-hand
side variables.}

\item{DF.Fit.Predict}{a data frame with training data and test data for the
estimation estage.}

\item{...}{Additional parameters passed to \link[stats]{glm}.}
}
\value{
Returns an object of class inheriting from "glm" (see
    \link[stats]{glm}).
}
\description{
Given the number of the equation that minimizes the selected error measure,
\code{fit_best} estimates the equation for each rolling sample and forecast
lead.
}
\keyword{internal}
