% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_predict.R
\name{fit_predict}
\alias{fit_predict}
\title{Fit and Predict Function}
\usage{
fit_predict(
  forecast.lead,
  eq.number,
  sample.number,
  glm.family,
  Y.name,
  all.equations.rhs,
  DF.Fit.Predict,
  ...
)
}
\arguments{
\item{forecast.lead}{A numeric value with the number of the forecast lead.}

\item{eq.number}{A numeric value with the number of the equation being
estimated.}

\item{sample.number}{A numeric value with the number of the rolling sample
being used in the estimation/prediction.}

\item{glm.family}{A description of the error distribution to be used in the
model. Inherits from \code{gears}.}

\item{Y.name}{Name of the y variable (variable to be forecasted) created by
\code{gears}.}

\item{all.equations.rhs}{Vector with all possible combinations of right-hand
side variables.}

\item{DF.Fit.Predict}{a data frame with training data and test data for the
estimation estage.}

\item{...}{Other parameters passed on the \link[stats]{glm}.}
}
\value{
Returns a numerical value giving the prediction for the selected
    equation, rolling sample, and forecast lead.
}
\description{
For each rolling sample, \code{fit_predict} uses the training data set to
fit a GLM. Then, it uses the testing data set to get the predictions.
}
\keyword{internal}
