% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_naive2.R
\name{forecast_naive2}
\alias{forecast_naive2}
\title{NAIVE2 Forecasts}
\usage{
forecast_naive2(ts.data, ts.frequency, forecast.horizon, alpha.level = 0.05)
}
\arguments{
\item{ts.data}{A numeric vector of time series observations (a ts object).}

\item{ts.frequency}{The frequency of a ts object.}

\item{forecast.horizon}{A numeric value with the length of the forecast lead.}

\item{alpha.level}{A numeric value with the alpha level to be used in the
test to detect seasonality. Default is 0.05.}
}
\value{
A ts object of forecasted values obtained using the Naive2 method.
}
\description{
Used to return the point forecasts using the Naive2 method. Code modified
from the M4 Competition's original code.
}
\examples{
forecast_naive2(
ts.data          = datasets::WWWusage,
ts.frequency          = stats::frequency(datasets::WWWusage),
forecast.horizon = 10,
alpha.level      = 0.05
)

forecast_naive2(
  ts.data          = datasets::AirPassengers,
  ts.frequency          = stats::frequency(datasets::AirPassengers),
  forecast.horizon = 15,
  alpha.level      = 0.05
)

forecast_naive2(
  ts.data          = datasets::EuStockMarkets[, "DAX"],
  ts.frequency          = stats::frequency(datasets::EuStockMarkets[, "DAX"]),
  forecast.horizon = 40,
  alpha.level      = 0.05
)
}
