% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package_help.R
\docType{package}
\name{gears-package}
\alias{gears-package}
\title{gears: GEARS Method for Univariate and Multivariate Time Series Forecasting}
\description{
Tools for forecasting univariate and multivariate time series
using the GEARS (GEneralized And Rolling Samples) Method.
}
