% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gears_optim.R
\name{gears_optim}
\alias{gears_optim}
\title{Optimization Algorithm for the GEARS Method for Time Series Forecasting}
\usage{
gears_optim(
  DATA,
  forecast.horizon,
  search.size.rs,
  search.number.rs,
  glm.family = c("gaussian", "binomial", "poisson", "Gamma", "quasi"),
  y.name = NULL,
  y.max.lags = NULL,
  x.names = NULL,
  x.max.lags = NULL,
  x.fixed.names = NULL,
  x.fixed.lags = NULL,
  x.interaction.names = NULL,
  x.interaction.lags = NULL,
  last.obs = NULL,
  use.intercept = c("both", "without", "with"),
  error.measure = c("mse", "mae", "mase", "smape", "list"),
  error.measure.list = NULL,
  betas.selection = c("last", "average", "both"),
  parallel.use = FALSE,
  DATA.insample,
  DATA.outsample,
  ...
)
}
\arguments{
\item{DATA}{A data frame or a univariate time series.}

\item{forecast.horizon}{Number of periods for forecasting.}

\item{search.size.rs}{A vector or a numeric value that indicates the
possible rolling sample sizes that the algorithm will search through.}

\item{search.number.rs}{A vector or a numeric value that indicates the
possible number of rolling sample that the algorithm will search through.}

\item{glm.family}{A description of the error distribution to be used in the
model. See \link[stats]{glm} for details.}

\item{y.name}{The name of the Y (left-hand side) variable. If NULL (default),
the function creates a temporary name.}

\item{y.max.lags}{A numeric value that gives the maximum number of lags of
the Y (left-hand side) variable (see Details). Can be NULL (default) if
the past values of the Y variable are not included in the model.}

\item{x.names}{List with names of the X (right-hand side) variables that
    have a maximum number of lags (see Details). Can be NULL (default) if
    univariate model or if your model does not have variables of this type.
#' @param x.max.lags List of numeric values that give the maximum number of lags
    of the X (right-hand side) variables. Can be NULL (default).}

\item{x.fixed.names}{List with names of the X (right-hand side) variables
that have a fixed number of lags (see Details).
Can be NULL (default) if univariate model or if your model does not have
variables of this type.}

\item{x.fixed.lags}{List of numeric values that give the fixed number of lags
of the variables in x.fixed.names. Can be NULL (default).}

\item{x.interaction.names}{List of character vectors with names of the
variables to be included as interaction terms (see Details).
Can be NULL (default) if your model does not have interaction terms.}

\item{x.interaction.lags}{List of numeric vectors with lags of the
variables to be included as interaction terms. List and numeric vectors
should have the same length as the ones in x.interaction.names.
Can be NULL (default) if your model does not have interaction terms.}

\item{last.obs}{Index number of the last observation to be considered.}

\item{use.intercept}{If \code{use.intercept == "both"} (default), the
function returns all possible model equations with and without intercept.
If \code{use.intercept == "without"}, then the function returns all
possible right-hand side equations without intercept. If
\code{use.intercept == "with"}, then only the right-hand side equations
with intercept are returned.}

\item{error.measure}{Error measure to be used when calculating the in-sample
prediction errors. Use \code{error.measure == "list"} if more than one
measure should be used. Then, add the list with measures to the
\code{error.measure.list} argument.}

\item{error.measure.list}{List with the names of the error measures to be
used when calculating the in-sample prediction errors. Only use this if
\code{error.measure == "list"}.}

\item{betas.selection}{If \code{betas.selection == "last"}, the
estimated coefficients from the last rolling sample will be used to
obtain the out-of-sample forecasts. If
\code{betas.selection == "average"}, then the average of the coefficients
from all rolling samples will be used.
If \code{betas.selection == "both"}, then two out-of-sample forecasts
will be estimated (one with \code{betas.selection == "last"} and
another with \code{betas.selection == "average"}).}

\item{parallel.use}{Boolean. Whether or not parallel computing should be
used. \code{gears_optim} uses the \code{\link[future.apply]} package
\url(https://github.com/HenrikBengtsson/future.apply)}

\item{DATA.insample}{Training data set. Either as a data frame object, or
a time series (ts) object.}

\item{DATA.outsample}{Testing data set. Either as a data frame object, or
a time series (ts) object.}

\item{...}{Further arguments passed to \link[stats]{glm}.}
}
\value{
A list containing the following
    elements:
    \describe{
     \item{fitted_best}{A list containing information about
         the best fitted model for each error measure (outer level of the
         list), forecast horizon (inner level 1), and each rolling sample
         (inner level 2).}
     \item{out_sample_point_forecasts}{A data frame containing point
         forecasts for each error measure and beta-selection method.}
     \item{insample_point_estimates}{List containing the out-of-sample point
         estimates for each rolling sample (row) and each forecast horizon
         (column).}
     \item{insample_mean_error_measure}{A data frame containing the mean
         prediction errors for each error measure and forecast horizon.}
     \item{best_equations}{A data frame containing information about the best
         equation fitted for each error measure and forecast horizon.}
    }
}
\description{
Finds the sample size and number of samples that minimizes the in-sample
(ex-ante) forecasting error.
}
