% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gears.R
\name{plot.gears}
\alias{plot.gears}
\title{Plot method for "gears" class}
\usage{
\method{plot}{gears}(
  x,
  col = 1,
  fcol = 4,
  ylim = NULL,
  main = NULL,
  xlab = "",
  ylab = "",
  type = "l",
  flty = 1,
  flwd = 2,
  ...
)
}
\arguments{
\item{x}{An object of class "gears".}

\item{col}{Color for the line with observed values. Default is 1 ("black").}

\item{fcol}{Color for the line with forecasts. Default is 4 ("blue").}

\item{ylim}{Limits for the Y-axis.}

\item{main}{Main title for the plot.}

\item{xlab}{Label for the X-axis.}

\item{ylab}{Label for the Y-axis.}

\item{type}{Line type for the line with observed values. See
\link[graphics]{plot} for more options.}

\item{flty}{Line type for the line with forecasts.}

\item{flwd}{Width of the line with forecasts.}

\item{...}{Other arguments passed to or from other methods.}
}
\description{
Plot method for "gears" class
}
