% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_errors.R
\name{prediction_errors}
\alias{prediction_errors}
\title{Returns the Forecast Errors for the GEARS Method.}
\usage{
prediction_errors(
  DATA,
  forecast.horizon,
  Y.name,
  total.equations,
  number.rs,
  DF.Fit.Predict,
  forecasts.gears,
  names.measures = c("mse", "mad", "smape", "mase", "owa")
)
}
\arguments{
\item{DATA}{A data frame or a univariate time series.}

\item{forecast.horizon}{Numeric value giving the number of periods for
forecasting.}

\item{Y.name}{Character string with the name of the Y (left-hand side)
variable.}

\item{total.equations}{A numeric value giving the number of equations to be
estimated (i.e., the number of equations from the output of
\link[gears]{all_models_rhs})}

\item{number.rs}{A numeric value giving the number of rolling samples.}

\item{DF.Fit.Predict}{A data frame with the training data and testing data.
It is the output of \link[gears]{create_DF_Fit_Predict}.}

\item{forecasts.gears}{A vector containing the forecasted values using the
GEARS method. It is the output of \code{prediction.gears} inside
\code{gears}.}

\item{names.measures}{Character vector with the names of the error measures
used in \link[gears]{error_measures}.}
}
\value{
Returns a list. First level of this list represents the different
    forecast leads. The table inside each level returns the equation/model's
    number (row) and the error measure (column), and the values inside the
    cells are the forecast errors.
}
\description{
Calculates the forecast errors using all methods in
\link[gears]{error_measures}: "MSE", "MAD", "sMAPE", "MASE", and "OWA".
}
\keyword{internal}
