% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_gears.R
\name{print.gears}
\alias{print.gears}
\title{Print method for "gears" class}
\usage{
\method{print}{gears}(x, ...)
}
\arguments{
\item{x}{An object of class "gears"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for "gears" class
}
