% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_test.R
\name{seasonality_test}
\alias{seasonality_test}
\title{Seasonality Test}
\usage{
seasonality_test(ts.data, ts.freq, alpha.level = 0.05)
}
\arguments{
\item{ts.data}{A ts object (i.e., a vector of time series observations).}

\item{ts.freq}{The frequency of a ts object.}

\item{alpha.level}{The alpha level to be used in the test to detect
seasonality. Default is 0.05.}
}
\value{
A boolean indicating whether or not the ts object has seasonality.
}
\description{
Used to determine whether a time series is seasonal. Code modified from the
M4 Competition's original code.
}
\examples{
seasonality_test(
  ts.data = datasets::WWWusage,
  ts.freq = frequency(datasets::WWWusage),
  alpha.level = 0.05
)

seasonality_test(
  ts.data = datasets::AirPassengers,
  ts.freq = frequency(datasets::AirPassengers),
  alpha.level = 0.05
)

seasonality_test(
  ts.data = datasets::EuStockMarkets[, "DAX"],
  ts.freq = frequency(datasets::EuStockMarkets[, "DAX"]),
  alpha.level = 0.05
)
}
