% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_gears.R
\name{summary.gears}
\alias{summary.gears}
\title{Summary method for "gears" class}
\usage{
\method{summary}{gears}(object, ...)
}
\arguments{
\item{object}{An object of class "gears"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Summary method for "gears" class
}
