% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_naive2.R
\name{forecast_naive2}
\alias{forecast_naive2}
\title{Title}
\usage{
forecast_naive2(ts.data, forecast.horizon, alpha.level = 0.05)
}
\arguments{
\item{ts.data}{A numeric vector of time series observations (a ts object).}

\item{forecast.horizon}{A numeric value with the length of the forecast lead.}

\item{alpha.level}{A numeric value with the alpha level to be used in the
test to detect seasonality. Default is 0.05.}
}
\value{
A ts object of forecasted values obtained using the Naive2 method.
}
\description{
Title
}
\examples{
forecast_naive2(
ts.data          = datasets::WWWusage,
forecast.horizon = 10,
alpha.level      = 0.05
)

forecast_naive2(
  ts.data          = datasets::AirPassengers,
  forecast.horizon = 15,
  alpha.level      = 0.05
)

forecast_naive2(
  ts.data          = datasets::EuStockMarkets[, "DAX"],
  forecast.horizon = 40,
  alpha.level      = 0.05
)
}
