% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gears_optim.R
\name{gears_optim}
\alias{gears_optim}
\title{Optimization Method for the GEARS Method for Time Series Forecasting}
\usage{
gears_optim(
  DATA,
  forecast.horizon,
  search.size.rs,
  search.number.rs,
  glm.family = c("gaussian", "binomial", "poisson", "Gamma", "quasi"),
  level = 95,
  details = FALSE,
  y.name = NULL,
  y.max.lags = NULL,
  x.names = NULL,
  x.max.lags = NULL,
  x.fixed.names = NULL,
  x.fixed.lags = NULL,
  x.interaction.names = NULL,
  x.interaction.lags = NULL,
  last.obs = NULL,
  use.intercept = c("both", "without", "with"),
  error.measure = c("mse", "mae", "mase", "smape", "owa"),
  betas.selection = c("last", "average", "both"),
  use.parallel = FALSE,
  num.cores = NULL,
  ...
)
}
\arguments{
\item{DATA}{A data frame or a univariate time series.}

\item{forecast.horizon}{Number of periods for forecasting.}

\item{search.size.rs}{Vector with the number of observations in the rolling
sample to be optimized.}

\item{search.number.rs}{Vector with the number of rolling samples to be
optimized.}

\item{glm.family}{A description of the error distribution to be used in the
model. See \link[stats]{glm} for details.}

\item{level}{Confidence level for prediction intervals. Numeric value
between 0 and 100.}

\item{details}{If \code{TRUE}, the function outputs the entire glm object for
all random samples for each forecast lead. To save memory, the default
is \code{FALSE}.}

\item{y.name}{The name of the Y (left-hand side) variable. If NULL (default),
the function creates a temporary name.}

\item{y.max.lags}{A numeric value that gives the maximum number of lags of
the Y (left-hand side) variable (see Details). Can be NULL (default) if
the past values of the Y variable are not included in the model.}

\item{x.names}{List with names of the X (right-hand side) variables that
have a maximum number of lags (see Details). Can be NULL (default) if
univariate model or if your model does not have variables of this type.}

\item{x.max.lags}{List of numeric values that give the maximum number of lags
of the X (right-hand side) variables. Can be NULL (default).}

\item{x.fixed.names}{List with names of the X (right-hand side) variables
that have a fixed number of lags (see Details).
Can be NULL (default) if univariate model or if your model does not have
variables of this type.}

\item{x.fixed.lags}{List of numeric values that give the fixed number of lags
of the variables in x.fixed.names. Can be NULL (default).}

\item{x.interaction.names}{List of character vectors with names of the
variables to be included as interaction terms (see Details).
Can be NULL (default) if your model does not have interaction terms.}

\item{x.interaction.lags}{List of numeric vectors with lags of the
variables to be included as interaction terms. List and numeric vectors
should have the same length as the ones in x.interaction.names.
Can be NULL (default) if your model does not have interaction terms.}

\item{last.obs}{Index number of the last observation to be considered.}

\item{use.intercept}{If \code{use.intercept == "both"} (default), the
function returns all possible model equations with and without intercept.
If \code{use.intercept == "without"}, then the function returns all
possible right-hand side equations without intercept. If
\code{use.intercept == "with"}, then only the right-hand side equations
with intercept are returned.}

\item{error.measure}{Error measure to be used when calculating the in-sample
prediction errors.}

\item{betas.selection}{If \code{betas.selection == "last"}, the
estimated coefficients from the last rolling sample will be used to
obtain the out-of-sample forecasts. If
\code{betas.selection == "average"}, then the average of the coefficients
from all rolling samples will be used.
If \code{betas.selection == "both"}, then two out-of-sample forecasts
will be estimated (one with \code{betas.selection == "last"} and
another with \code{betas.selection == "average"}).}

\item{use.parallel}{If parallel computing should be used. Default is
\code{FALSE}.}

\item{num.cores}{Number of cores to use if parallel computing is used.
If \code{use.parallel == TRUE}, then the default is
\code{detectCores() - 1}. For more details, see
\link[parallel]{detectCores}.}

\item{...}{Further arguments passed to \link[stats]{glm}.}
}
\value{
A data frame with four columns: "size.rs", "number.rs", "intercept",
   and the last one named after the selected error measure. The last column
   returns the minimum value for the selected error measure, and the columns
   "size.rs" and "number.rs" return the sample sizes and the number of
   rolling samples used to get this minimum. The column "intercept" gives
   information on whether the intercept was used ("with"), or not
   ("without").
}
\description{
This algorithm will search for the "best" size of the rolling sample and the
"best" number of rolling samples. "Best" in the sense of minimizing the
selected error measure.
}
\details{
If \code{y.max.lags} equals to a number, then all lags of
    \code{y.name} up to this number (and starting at 0) will be included
    in the list of variables to enter the right-hand side of the model
    equations. For example, if \code{y.max.lags = 2}, then
    \ifelse{html}{\out{Y<sub>t</sub>, Y<sub>t-1</sub>, Y<sub>t-2</sub>
    }}{\eqn{Y_t, Y_{t-1},Y_{t-2}}} will be included in the list of variables
    to enter the right-hand side of the equation.
}
\examples{
# Univariate Time Series Forecasting - Data of class 'ts'.
gears_optim(
  DATA                = datasets::WWWusage,
  forecast.horizon    = 12,
  search.size.rs      = c(20, 30),
  search.number.rs    = c(10, 12),
  level               = 95,
  details             = FALSE,
  glm.family          = "quasi",
  y.name              = NULL,
  y.max.lags          = 2,
  x.names             = NULL,
  x.max.lags          = NULL,
  x.fixed.names       = NULL,
  x.fixed.lags        = NULL,
  x.interaction.names = NULL,
  x.interaction.lags  = NULL,
  last.obs            = length(datasets::WWWusage),
  use.intercept       = "both",
  error.measure       = "mse",
  betas.selection     = "last",
  use.parallel        = FALSE,
   num.cores          = NULL
)

gears_optim(
  DATA                = datasets::WWWusage,
  forecast.horizon    = 12,
  search.size.rs      = c(20, 30),
  search.number.rs    = c(10, 12),
  level               = 95,
  details             = FALSE,
  glm.family          = "quasi",
  y.name              = NULL,
  y.max.lags          = 2,
  x.names             = NULL,
  x.max.lags          = NULL,
  x.fixed.names       = NULL,
  x.fixed.lags        = NULL,
  x.interaction.names = NULL,
  x.interaction.lags  = NULL,
  last.obs            = length(datasets::WWWusage),
  use.intercept       = "both",
  error.measure       = "mse",
  betas.selection     = "both",
  use.parallel        = FALSE,
  num.cores           = NULL
)

gears_optim(
  DATA                = datasets::WWWusage,
  forecast.horizon    = 12,
  search.size.rs      = c(20, 30),
  search.number.rs    = c(10, 12),
  level               = 95,
  details             = FALSE,
  glm.family          = "quasi",
  y.name              = NULL,
  y.max.lags          = 2,
  x.names             = NULL,
  x.max.lags          = NULL,
  x.fixed.names       = NULL,
  x.fixed.lags        = NULL,
  x.interaction.names = NULL,
  x.interaction.lags  = NULL,
  last.obs            = length(datasets::WWWusage),
  use.intercept       = "with",
  error.measure       = "mse",
  betas.selection     = "both",
  use.parallel        = FALSE,
  num.cores           = NULL
)

# With Parallel Computing
# Univariate Time Series Forecasting - Data of class 'ts'.
gears_optim(
  DATA                = datasets::WWWusage,
  forecast.horizon    = 12,
  search.size.rs      = c(20, 30),
  search.number.rs    = c(10, 12),
  level               = 95,
  details             = FALSE,
  glm.family          = "quasi",
  y.name              = NULL,
  y.max.lags          = 2,
  x.names             = NULL,
  x.max.lags          = NULL,
  x.fixed.names       = NULL,
  x.fixed.lags        = NULL,
  x.interaction.names = NULL,
  x.interaction.lags  = NULL,
  last.obs            = length(datasets::WWWusage),
  use.intercept       = "both",
  error.measure       = "mse",
  betas.selection     = "last",
  use.parallel        = TRUE,
  num.cores          = NULL
)

}
\author{
Gustavo Varela-Alvarenga
}
\keyword{glm}
\keyword{optimization}
\keyword{ts}
